pyproject.toml
[project]
name = "recommender-system"
version = "0.1.0"
description = "Recommender System - Clean Architecture (Lab 1)"
authors = [
    {name = "lolovlad",email = "lolovlad33@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "dvc[s3] (>=3.66.1,<4.0.0)",
    "boto3 (>=1.41.0,<1.41.6)",
    "fastapi (>=0.116.1,<1.0.0)",
    "uvicorn (>=0.35.0,<1.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "onnx (>=1.20.1,<2.0.0)",
    "onnxruntime (>=1.24.4,<2.0.0)",
    "skl2onnx (>=1.20.0,<2.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "recommender_system", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<1.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
